+++ RELEASE-NOTES.md
@@ -8,6 +8,7 @@
 * #20 Split healthchecks code to a new "bootique-metrics-healthchecks"
 * #21 Upgrade to Dropwizard metrics 3.2.5
 * #22 HealthCheckRegistry filtering
+* #23 Nagios-like health-check statuses
 
 ## 0.10